requirements.txt
django==2.2
flake8==3.8.3
djangorestframework==3.10.0
yapf==0.30.0
pylint==2.6.0
djangorestframework-simplejwt==4.4.0
pyyaml==5.3.1
mock==4.0.2
factory-boy==3.0.1
freezegun==1.0.0
mysqlclient==2.0.1
django-environ==0.4.5
responses==0.12.0
django-phonenumber-field==2.0.0